requirements.txt
boto3
pylint
pytest
black
pyyaml

